pyproject.toml
[tool.poetry]
name = "demowebshop-tests"
version = "0.1.0"
description = "Automated tests for Demowebshop"
authors = ["TommyMadness"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
requests = ">=2.32.3,<3.0.0"
pytest = ">=8.3.5,<9.0.0"
jsonschema = ">=4.23.0,<5.0.0"
selene = "==2.0.0rc9"
allure-pytest = ">=2.14.0,<3.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
